+++ database/migrations/2018_03_29_123639_create_bill_of_materials_table.php
@@ -29,11 +29,11 @@ class CreateBillOfMaterialsTable extends Migration
             $table->integer('part_changes')->nullable();
             $table->integer('material_id')->nullable();
             $table->string('part_serial_number')->nullable();
-            $table->integer('quantity_part_thickness')->nullable();
-            $table->integer('quantity_part_width')->nullable();
-            $table->integer('quantity_part_length')->nullable();
-            $table->integer('quantity_part_diameter')->nullable();
-            $table->integer('quantity_part_weight')->nullable();
+            $table->decimal('quantity_part_thickness')->nullable();
+            $table->decimal('quantity_part_width')->nullable();
+            $table->decimal('quantity_part_length')->nullable();
+            $table->decimal('quantity_part_diameter')->nullable();
+            $table->decimal('quantity_part_weight')->nullable();
             $table->string('dimension_uom')->nullable();
             $table->string('remark')->nullable();
             $table->boolean('is_approved')->nullable()->default(false);
